+++ packages/form-handler-bundle/src/ViewModel/FormAwareInterface.php
@@ -18,13 +18,9 @@ interface FormAwareInterface
 {
     public function setForm(FormInterface $form): void;
 
-    public function getForm(): FormInterface;
+    public function getForm(): ?FormInterface;
 
-    public function setIsSuccess(bool $isSuccess): void;
-
-    public function getIsSuccess(): bool;
-
-    public function getFormView(): FormView;
+    public function getFormView(): ?FormView;
 
     public function formIsValid(): bool;
 }